Fixed path names to get b26_run_gui to work.
Deleted a line from Jan's fpga code in Ni_9263_polarization_controller.py. This was causing a bug when running a script iterator.

Aaron also "undid a patch about for a bug that was fixed by adding an init file to the top level of the b26_toolkit directory." That's in core/scripts.py and tools/export_default_v2.py