plots.py: removed axes.hold() functions (removed in new version of matplotlib). did the same in Rabi, Hahn echo and ESR pulse sequence files.
also added 'safety' code to the pulse sequence files (including the pulsed experiment base script) that makes srue the mw_switch is OFF when experiments aren't running, turns on when they are, and turns on IQ modulation when necessary.

ServoScan.py: temporarily removed the piezo as an instrument on this script because it wasn't connecting properly and we needed to move forward. It'll throw an error if you try to run the script as it is now.